# src/app/api/v1/[[...route]]/openapi.yaml
openapi: 3.1.0
info:
  title: WorkoraJobs Automation API Layer
  version: 1.0.0
  description: |
    Production-grade backend automation layer endpoints for WorkoraJobs.
    Provides structural integration interfaces for n8n, AI agents, mobile clients, and developer portals.
    
    ### Authentication
    All API endpoints require one of the following credentials:
    1. **API Key Authentication**: Injected via header as `x-api-key: <key>`.
    2. **Bearer JWT Token**: Injected via header as `Authorization: Bearer <JWT>`.

    ### Rate Limits
    - Rate limits are managed via Redis and allow up to `100 requests / minute` per client IP.
    - Rate limit headers are automatically returned in responses.
paths:
...
  /jobs:
    get:
      summary: List all jobs
      responses:
        '200':
          description: A list of active job listings
    post:
      summary: Create a job listing
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                location:
                  type: string
                salary:
                  type: integer
                type:
                  type: string
                  enum: [FULL_TIME, PART_TIME, CONTRACT, INTERNSHIP, FREELANCE]
                companyId:
                  type: string
                postedById:
                  type: string
              required: [title, description, type, companyId, postedById]
      responses:
        '201':
          description: Job created successfully
        '400':
          description: Validation error
        '409':
          description: Duplicate listing detected
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT